+++ src/routes/(v2)/v2/enum-proxy/schema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+const fishes = ['trout', 'tuna', 'shark'] as const;
+
+export const schema = z.object({
+	fish: z.enum(fishes),
+	moreFish: z.enum(fishes)
+});